src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/text/primitive/SinglePrimitiveShortTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.primitive;

import com.github.brunomndantas.jscrapper.support.parser.single.text.SingleTextParser;
import org.openqa.selenium.WebElement;

public class SinglePrimitiveShortTextParser extends SingleTextParser {

    public SinglePrimitiveShortTextParser() {
        super((short)0);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Short.parseShort(element.getText());
    }

}